ygen
Module IR

The ir module: functions for building function ir

Modules

ir

Stores all ir nodes and the ir trait

parser

This module contains the logic required for parsing ygen-ir files It defines the Lexer and the Parser Additionaly it provides an easy to use function which is a wrapper around the lexing, parsing unit

Structs

Block

A basic block: stores ir of a specific area of a function

BlockId

Describes the block

Const

A constant

FuncId

A function id (describes arguments and the name)

Function

A ir function with a known variable and arg size and count

FunctionType

Stores the function type

Module

The Module

Var

A variable

Enums

Type

Stores a type and a value of that type

TypeMetadata

Stores type metadata (just the type without data)

VerifyError

An error which stores if an ir node is invalid

Functions

Block

Creates an new block

FnTy

Creates a new function type

Func

Creates a new Function

Module

Creates a new module

Var

Creates a new variable