ygen::IR
Enum Type

pub enum Type {
    u8(u8),
    u16(u16),
    u32(u32),
    u64(u64),
    i8(i8),
    i16(i16),
    i32(i32),
    i64(i64),
    ptr(i64),
    Void,
    f32(f32),
    f64(f64),
}

Stores a type and a value of that type

If you want an empty Type consider using TypeMetadata

Variants

u8(u8)

Just an u8 with a value

u16(u16)

Just an u16 with a value

u32(u32)

Just an u32 with a value

u64(u64)

Just an u64 with a value

i8(i8)

Just an i8 with a value

i16(i16)

Just an i16 with a value

i32(i32)

Just an i32 with a value

i64(i64)

Just an i64 with a value

ptr(i64)

64Bit pointer

Void

Notype

f32(f32)

A f32

f64(f64)

A f64

Implementations

impl Type

pub fn val(&self) -> f64

Returns the inner value

pub fn from_int(ty: TypeMetadata, value: f64) -> Self

puts the intenger into a type respecting the type metadata