ygen::IR::ir
Trait BuildBrCond

pub trait BuildBrCond<T, U, Z> {
    // Required method
    fn BuildBrCond(&mut self, val: T, iftrue: U, iffalse: Z);
}

This trait is used for building br condition nodes

Required Methods

fn BuildBrCond(&mut self, val: T, iftrue: U, iffalse: Z)

Builds a br condition node

Jumps to iftrue if the value is not 0 else to iffalse