ygen::IR::ir
Trait BuildCast

pub trait BuildCast<T, U> {
    // Required method
    fn BuildCast(&mut self, value: T, ty: U) -> Var;
}

Trait for the cast instruction Used for overloading the BuildCast function

Required Methods

fn BuildCast(&mut self, value: T, ty: U) -> Var

builds an cast to form one variable into another type