Trait ygen::IR::ir::Ir

pub trait Ir: Debug + Any {

    // Required methods
    fn dump(&self) -> String;
    fn dumpColored(&self, profile: ColorProfile) -> String;
    fn as_any(&self) -> &dyn Any;
    fn verify(&self, _: FunctionType) -> Result<(), VerifyError>;
    fn clone_box(&self) -> Box<dyn Ir>;
    fn compile(&self, registry: &mut TargetBackendDescr, module: &mut Module);
    fn compile_dir(
        &self,
        compiler: &mut IrCodeGenHelper,
        block: &Block,
        module: &mut Module,
    );
    fn maybe_inline(
        &self,
        const_values: &HashMap<String, Type>,
    ) -> Option<Box<dyn Ir>>;
    fn eval(&self) -> Option<Box<dyn Ir>>;
    fn inputs(&self) -> Vec<Var>;
    fn output(&self) -> Option<Var>;

    // Provided methods
    fn uses(&self, _: &Var) -> bool { ... }
    fn is(&self, other: &Box<dyn Ir>) -> bool { ... }

}

The ir trait

Required Methods

fn dump(&self) -> String

Returns the ir node as his textual representation

fn dumpColored(&self, profile: ColorProfile) -> String

Returns the ir node as his textual representation with colors

fn as_any(&self) -> &dyn Any

Turns the ir node to an any

fn verify(&self, _: FunctionType) -> Result<(), VerifyError>

verifys the instruction (used for return instruction) based on the return type

fn clone_box(&self) -> Box<dyn Ir>

Clones the node into a box of Box<dyn Ir>

fn compile(&self, registry: &mut TargetBackendDescr, module: &mut Module)

Compiles the node based on the given target

fn compile_dir( &self, compiler: &mut IrCodeGenHelper, block: &Block, module: &mut Module, )

Compiles the node with node information to the given target

fn maybe_inline( &self, const_values: &HashMap<String, Type>, ) -> Option<Box<dyn Ir>>

inlines the variables to the types if possible

fn eval(&self) -> Option<Box<dyn Ir>>

evaluteas the node

fn inputs(&self) -> Vec<Var>

returns the vars used by the node as input

fn output(&self) -> Option<Var>

returns the output var

Provided Methods

fn uses(&self, _: &Var) -> bool

Returns if the node uses the variable

fn is(&self, other: &Box<dyn Ir>) -> bool

checks if the node is equal to the other node (used for the implementation of Eq)