Struct ygen::IR::Function

pub struct Function {
    pub ty: FunctionType,
    /* private fields */
}

A ir function with a known variable and arg size and count

Fields
ty: FunctionType

The function type

Implementations

impl Function

pub fn new(name: String, ty: FunctionType) -> Self

Creates a new Function

pub fn extrn(&mut self)

Sets that the function is externally visible (same as: extern "C")

pub fn import(&mut self)

Sets that the function is imported from another object file (same as: extern "C" fn abc(i32, i32) -> i32;)

pub fn private(&mut self)

Sets that the function is only internally visible (same as a normal function)

pub fn addBlock(&mut self, name: &str) -> BlockId

Adds a new block to the function

pub fn dump(&self) -> String

Emits the Ir of the function into a string

pub fn dumpColored(&self, profile: ColorProfile) -> String

Emits the Ir of the function into an colored string

pub fn verify(&self) -> Result<(), VerifyError>

Verifys if the function and all of its blocks are correct:

Checks if the return type is the actual specified return type of the function

Checks all ir nodes

pub fn runPassMngr(&mut self, mngr: &PassManager)

Runs the pass manager over all blocks

pub fn id(&self) -> FuncId

Constructs an id for the function

impl Function

pub fn BuildAlloca(&mut self, ty: TypeMetadata) -> Var

Builds an stack allocation (the out var is the pointer to the allocated stack region)

impl Function

pub fn BuildLoad(&mut self, ptr: Var, ty: TypeMetadata) -> Var

the load instruction loads an value from an pointer into a normal variable

impl Function

pub fn BuildDebug(&mut self, line: i64, coloumn: i64, file: PathBuf)

Sets the source location for debugging (all of the ir nodes will respond to the location till an new location is set)

impl Function

pub fn BuildPhi( &mut self, typ: TypeMetadata, recipients: Vec<(&Block, Var)>, ) -> Var

Builds the phi node which recives variables from different blocks

impl Function

pub fn BuildSwitch( &mut self, source: Var, default: &BlockId, cases: HashMap<Type, &BlockId>, )

Builds an switch statement

impl Function

pub fn BuildNeg(&mut self, var: Var) -> Var

Builds a negate (negates the content of the variable)

impl Function

pub fn BuildGetelemptr(&mut self, var: Var, index: Var, ty: TypeMetadata) -> Var

Builds the getelemptr node. It is used for example in this c code:

int test(int a[5]) {
    return a[1];
}

It would be compiled to this:

define i32 @test(ptr %0) {
  entry:
    %const0 = i32 1
    %1 = getelemptr ptr %0, i32 %const0, i32
    ret i32 %1
}