Struct ygen::IR::IRBuilder

pub struct IRBuilder<'a> { /* private fields */ }

IRBuilder: used for building the ir of a function

Implementations

impl<'a> IRBuilder<'a>

pub fn new() -> Self

Creates an new ir builder

pub fn positionAtEnd(&mut self, block: &'a mut Block)

Positions the block at the end of the blocks list

pub fn positionAtStart(&mut self, block: &'a mut Block)

Positions the block at the start of the blocks list

pub fn getLastBlock(&mut self) -> Option<&Block>

Returns the last block of the builder