Module ygen::Support

Other utilites like:

Cli args

String coloring and padding

Structs

Cli

A command line parser

Color

A simple color

ColorEncoder

Colore encoder: Encodes a string by a specifed format E.g: Hi I am blueHi I am green For that use the encode function

ColorProfile

The user color profile is used to store how ir elements are colored For example instructions are red

Error

An Error: an Error struct can store Error informations

SrcMngr

The Source Manager: easly chars out of a file

TokenMgr

The token manager

TypeSwitch

Advanced type switching

Enums

ColorClass

The color markup class

Traits

Colorize

Clorize trait: Colorizes a string

Pad

A trait to pad strings

PrintErrorAndExit

Prints the error and exits the process If no error occured this function just returns