ygen::Support
Trait PrintErrorAndExit

pub trait PrintErrorAndExit {
    // Required method
    fn print(&self);
}

Prints the error and exits the process If no error occured this function just returns

Required Methods

fn print(&self)

Prints the error and exits the process If no error occured this function just returns