ygen::Target::x64
Enum X64Reg

pub enum X64Reg {

    Rax,
    Eax,
    Ax,
    Al,
    Rbx,
    Ebx,
    Bx,
    Bl,
    Rcx,
    Ecx,
    Cx,
    Cl,
    Rdx,
    Edx,
    Dx,
    Dl,
    Rsi,
    Esi,
    Si,
    Sil,
    Rdi,
    Edi,
    Di,
    Dil,
    Rsp,
    Esp,
    Sp,
    Spl,
    Rbp,
    Ebp,
    Bp,
    Bpl,
    R8,
    R8d,
    R8w,
    R8b,
    R9,
    R9d,
    R9w,
    R9b,
    R10,
    R10d,
    R10w,
    R10b,
    R11,
    R11d,
    R11w,
    R11b,
    R12,
    R12d,
    R12w,
    R12b,
    R13,
    R13d,
    R13w,
    R13b,
    R14,
    R14d,
    R14w,
    R14b,
    R15,
    R15d,
    R15w,
    R15b,
    Xmm0,
    Xmm1,
    Xmm2,
    Xmm3,
    Xmm4,
    Xmm5,
    Xmm6,
    Xmm7,
    Xmm8,
    Xmm9,
    Xmm10,
    Xmm11,
    Xmm12,
    Xmm13,
    Xmm14,
    Xmm15,

}

A x64 register

Variants

Rax

Eax

Ax

Al

Rbx

Ebx

Bx

Bl

Rcx

Ecx

Cx

Cl

Rdx

Edx

Dx

Dl

Rsi

Esi

Si

Sil

Rdi

Edi

Di

Dil

Rsp

Esp

Sp

Spl

Rbp

Ebp

Bp

Bpl

R8

R8d

R8w

R8b

R9

R9d

R9w

R9b

R10

R10d

R10w

R10b

R11

R11d

R11w

R11b

R12

R12d

R12w

R12b

R13

R13d

R13w

R13b

R14

R14d

R14w

R14b

R15

R15d

R15w

R15b

Xmm0

Xmm1

Xmm2

Xmm3

Xmm4

Xmm5

Xmm6

Xmm7

Xmm8

Xmm9

Xmm10

Xmm11

Xmm12

Xmm13

Xmm14

Xmm15

Implementations

impl X64Reg

pub fn parse(string: String) -> Option<Self>

Parses the string to an register (Returns none if it’s invalid)

pub fn extended(&self) -> bool

Returns if the reg is in the extendet region (r8->r15)

pub fn sub64(&self) -> X64Reg

The sub 64 bit wide variant of the register

pub fn sub32(&self) -> X64Reg

The sub 32 bit wide variant of the register

pub fn sub16(&self) -> X64Reg

The sub 16 bit wide variant of the register

pub fn sub8(&self) -> X64Reg

The sub8 bit wide variant of the register

pub fn sub_ty(&self, ty: TypeMetadata) -> X64Reg

gets the subvariant based on the type

pub fn is_gr64(&self) -> bool

Is the register (or better the subvariant) 64 bit wide?

pub fn is_gr32(&self) -> bool

Is the register (or better the subvariant) 32 bit wide?

pub fn is_gr16(&self) -> bool

Is the register (or better the subvariant) 16 bit wide?

pub fn is_gr8(&self) -> bool

Is the register (or better the subvariant) 8 bit wide?

pub fn is_xmm(&self) -> bool

Is the register a xmm register?

pub fn callee_saved(&self, call: CallConv) -> bool

Returns if the register is callee saved on the given calling convention