Struct ygen::Target::x64::instr::X64MCInstr

pub struct X64MCInstr {
    pub mnemonic: Mnemonic,
    pub op1: Option<Operand>,
    pub op2: Option<Operand>,
    /* private fields */
}

The target instruction

Fields
mnemonic: Mnemonic

The mnemonic to use

op1: Option<Operand>

First operand

op2: Option<Operand>

Second operand

Implementations

impl X64MCInstr

pub fn with0(mne: Mnemonic) -> Self

Creates the instruction with 0 operands

pub fn make_far(&self) -> Self

Makes the instruction use far calls (for call and so on)

pub fn with1(mne: Mnemonic, op: Operand) -> Self

Creates the instruction with 1 operand

pub fn with2(mne: Mnemonic, op1: Operand, op2: Operand) -> Self

Creates the instruction with 2 operands

pub fn encode(&self) -> Result<(Vec<u8>, Option<Link>), Box<dyn Error>>

Encodes the instruction (some will say compiles)

pub fn verify(&self) -> Result<(), InstrEncodingError>

Verifys the instruction (like checking the right opcodes etc.)

pub fn compile(&self) -> Result<Vec<u8>, Box<dyn Error>>

Does the same as the encode function just for naming pourpuses

pub fn to_string(&self) -> String

Returns the instruction as assembly representation

pub fn color(&self, profile: ColorProfile) -> String

emits the instruction as one colored string

pub fn empty(&self) -> bool

returns if the instruction is empty like mov rsi, rsi

impl X64MCInstr

pub fn is_add(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_adc(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_and(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_or(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_xor(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sub(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_neg(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmp(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_lea(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mov(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movzx(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_push(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_pop(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ret(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_imul(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mul(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_idiv(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_div(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_call(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_jmp(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_jne(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_je(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_endbr64(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sete(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setne(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setg(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setl(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setge(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setle(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmove(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmovne(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sal(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_shr(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sar(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movq(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movsd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movups(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movupd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_addss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_addsd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_divss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_divsd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mulss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mulsd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_subss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_subsd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ucomiss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ucomisd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtss2si(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsd2si(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtss2sd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsd2ss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsi2ss(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsi2sd(&self) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_add1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_adc1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_and1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_or1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_xor1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sub1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_neg1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmp1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_lea1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mov1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movzx1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_push1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_pop1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_imul1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mul1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_idiv1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sete1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setne1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setg1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setl1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setge1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_setle1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmove1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmovne1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sal1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_shr1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movq1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movsd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movups1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movupd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_addss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_addsd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_divss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_divsd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mulss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mulsd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_subss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_subsd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ucomiss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ucomisd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtss2si1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsd2si1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtss2sd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsd2ss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsi2ss1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsi2sd1(&self, op1: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_add2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_adc2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_and2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_or2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_xor2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sub2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmp2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_lea2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mov2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movzx2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmove2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cmovne2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_sal2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_shr2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movq2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movsd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movups2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_movupd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_addss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_addsd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_divss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_divsd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mulss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_mulsd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_subss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_subsd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ucomiss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_ucomisd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtss2si2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsd2si2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtss2sd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsd2ss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsi2ss2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_cvtsi2sd2(&self, op1: &Operand, op2: &Operand) -> bool

[AUTO GENERATED] checks if the instruction is the same as the other

impl X64MCInstr

pub fn is_op1_reg(&self) -> bool

Checks if the first operand is a register

pub fn is_op1_mem(&self) -> bool

Checks if the first operand is a memory displacment

pub fn is_op2_reg(&self) -> bool

Checks if the second operand is a register

pub fn is_op2_mem(&self) -> bool

Checks if the second operand is a memory displacment