ygen
Module CodeGen

Shared code generation classes (mainly used for register based architectures like x86_64, aarch64, …)

Structs

Allocator

The allocation functions for the CompilationHelper

CompilationHelper

helps with compilation

IrCodeGenArea

The instructions for a single ir node

IrCodeGenHelper

CompilationHelper but with node metadata

MCDocInstr

a doc comment in the generated assembly code

MCSettings

Machine specific settings, like the lowering function prolog, epilog functions, etc

MachineCallingConvention

A more machine specifc calling convention (Just a wrapper around the normal calling convention but with some pretty handy functions)

MachineInstr

a low level instruction which is portable over platforms

RegVec

A register vector. Used for getting platform specific registers

Enums

ConstImmRules

handeles how constant imms are handeled (wether creating a const or just an instr op)

MachineMnemonic

The mnemonic to use

MachineOperand

a low level operand which is portable over platforms

Reg

A shared enum for registers

Traits

MCInstr

a platform specifc instruction