ygen::IR
Module ir

Stores all ir nodes and the ir trait

Structs

Add

An Ir node

Alloca

An Ir node

And

An Ir node

Assign

The assign ir node

Br

An Ir node

BrCond

An Ir node

Call

Call ir node

Cast

An Ir node

Cmp

The cmp node is used to compare values

DebugNode

A node which startes a debugging line programm

Div

An Ir node

GetElemPtr

extracts an element out of a vector

Load

An Ir node

Mul

An Ir node

Neg

An Ir node

Or

An Ir node

Phi

The phi node which is used to influence the register allocator

Rem

An Ir node

Return

An Ir node

Select

The select ir node

Shl

An Ir node

Shr

An Ir node

Store

An Ir node

Sub

An Ir node

Switch

The switch node is used to switch

Xor

An Ir node

Enums

CmpMode

The “compare mode” (e.g: ls is equal to rs)

IROperand

an operand for ir nodes

Traits

BuildAdd

Used for overloading the build function

BuildAnd

Used for overloading the build function

BuildAssign

Trait used for overloading the BuildAssign function

BuildBr

This trait is used for building br nodes

BuildBrCond

This trait is used for building br condition nodes

BuildCall

Trait for the call instruction Used for overloading the BuildCall function

BuildCast

Trait for the cast instruction Used for overloading the BuildCast function

BuildCmp

The trait BuildCmp is used to build the cmp node

BuildDiv

Used for overloading the build function

BuildMul

Used for overloading the build function

BuildOr

Used for overloading the build function

BuildRem

Used for overloading the build function

BuildReturn

Trait for the return instruction Used for overloading the BuildRet function

BuildSelect

This trait is used to build the select node

BuildShl

Used for overloading the build function

BuildShr

Used for overloading the build function

BuildStore

The BuildStore trait is used for overloading the BuildStore method

BuildSub

Used for overloading the build function

BuildXor

Used for overloading the build function

EvalOptVisitor

A trait used for constant propagination

Ir

The ir trait

IsNode

checks if the node is another node

Replace

Used for sus workaround to replace current ir node