Module Ygen::IR::ir

Stores all ir nodes and the ir trait

Structs

Add

An Ir node

And

An Ir node

Cast

An Ir node

ConstAssign

An Ir node

Or

An Ir node

Sub

An Ir node

Xor

An Ir node

Traits

BuildAdd

Used for overloading the build function

BuildAnd

Used for overloading the build function

BuildCast

Trait for the cast instruction Used for overloading the BuildCast function

BuildOr

Used for overloading the build function

BuildReturn

Trait for the return instruction Used for overloading the BuildRet function

BuildSub

Used for overloading the build function

BuildXor

Used for overloading the build function

Replace

Used for sus workaround to replace current ir node